Update BundleRetrieveProvider to handle code paths that point to references
Prior to this commit, BundleRetrieveProvider.retrieve() would throw an IllegalArgumentException that stated, "Unable to extract codes from fhirType Reference." when `codePath` pointed to a choice of data type element that held a `Reference` (e.g. MedicationAdministration.medication[x]). This commit makes BundleRetrieveProvider resilient to the case where the code path points to a `Reference`.